test: Provide clean configs for test npm instance
Use `user.npmrc` and `global.npmrc` in the current directory as
replacements for `~/.npmrc` and `$PREFIX/etc/npmrc` respectively. The
user can optionally provide these files as part of their test
environment; npm will treat them as empty if they are not present.

Since no host system configuration is loaded, add an email address to
the default config because it is a mandatory field for `npm publish`
based tests.